+++ src/components/index.js
@@ -1,0 +1,9 @@
+export * from './Aside';
+export * from './CreateUser';
+export * from './Content';
+export * from './Main';
+export * from './Tabs';
+export * from './Task';
+export * from './TaskList';
+export * from './UserList';
+
