+++ src/pages/Gallery/gallery.scss
