+++ src/pages/Gallery/index.js
