+++ src/parts/Footer/index.js
